models/requirements.txt
autogluon.tabular>=1.0
yfinance>=0.2.31
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
